Fix Probe Circle
Probe Circle commands were not stopping after the probe was triggered.  To fix this, I do an immediate_halt.  then if probe was detected, I reset the halt because that should have been what triggered the halt condition.

This could be better, but I don't know how else to stop the segments of the circle command from continuing after the probe.